[INTERNAL][I] Use toolbar project object for UI components
Adjusts the UI logic related to the toolbar and its contained components
to use the project object passed to the toolbar during its
initialization. This allows the UI components to always use the correct
project object as the reference point when creating new UI components
like dialogs.

Subsequently adjusts the dialog utility classes to accept a project
object as an additional parameter instead of always using the project
object contained in the plugin context.

As DialogUtils.showConfirmationDialog(...) is still used by components
that currently can't access a project object (besides the one held in
the plugin context), the method is kept as oldShowConfirmationDialog.
The method is marked as deprecated and will be removed once all required
components have access to a project object.